requirements.txt
google-cloud-storage>=2.0.0
google-cloud-bigquery>=2.0.0
requests>=2.25.0
python-dotenv>=0.21.0
